refactor: unify unicode helper modules
