+++ core/unicode_decode.py
@@ -1,50 +1,21 @@
 from __future__ import annotations
 
-"""Lightweight Unicode decoding helpers."""
+"""Deprecated: use :mod:`core.unicode` instead."""
 
-import logging
+import warnings
 from typing import Any
 
-logger = logging.getLogger(__name__)
+from .unicode import safe_unicode_decode as _safe_unicode_decode
 
-
-def _remove_invalid_surrogates(text: str) -> str:
-    """Return ``text`` without lone surrogate code points."""
-    out = []
-    i = 0
-    while i < len(text):
-        ch = text[i]
-        code = ord(ch)
-        if 0xD800 <= code <= 0xDBFF:
-            if i + 1 < len(text) and 0xDC00 <= ord(text[i + 1]) <= 0xDFFF:
-                out.append(ch)
-                out.append(text[i + 1])
-                i += 2
-                continue
-            i += 1
-            continue
-        if 0xDC00 <= code <= 0xDFFF:
-            i += 1
-            continue
-        out.append(ch)
-        i += 1
-    return "".join(out)
+warnings.warn(
+    "core.unicode_decode is deprecated; use core.unicode.safe_unicode_decode",
+    DeprecationWarning,
+    stacklevel=2,
+)
 
 
 def safe_unicode_decode(data: Any, encoding: str = "utf-8") -> str:
-    """Decode ``data`` safely removing invalid surrogates."""
-    if isinstance(data, str):
-        return _remove_invalid_surrogates(data)
-
-    try:
-        text = data.decode(encoding, errors="surrogatepass")
-    except UnicodeError as exc:
-        logger.warning("Primary decode failed: %s", exc)
-        try:
-            text = data.decode(encoding, errors="replace")
-        except Exception:
-            text = data.decode("utf-8", errors="ignore")
-    return _remove_invalid_surrogates(text)
+    return _safe_unicode_decode(data, encoding)
 
 
 __all__ = ["safe_unicode_decode"]
