+++ services/data_processing/file_handler.py
@@ -9,8 +9,7 @@ from config.dynamic_config import dynamic_config
 from config.constants import DEFAULT_CHUNK_SIZE
 from core.performance import get_performance_monitor
 from core.protocols import ConfigurationProtocol
-from core.unicode import process_large_csv_content
-from core.unicode_utils import sanitize_for_utf8
+from core.unicode import process_large_csv_content, sanitize_for_utf8
 from security.unicode_security_processor import sanitize_dataframe
 from services.data_processing.core.exceptions import (
     FileProcessingError,
